fix some warnings

+++ src/cursor.rs
@@ -103,7 +103,7 @@ pub fn pointer_cursor(
     // check if the cursor is inside the window and get its position
     if let Some(screen_pos) = wnd.cursor_position() {
         // get the size of the window
-        let window_size = Vec2::new(wnd.width() as f32, wnd.height() as f32);
+        let window_size = Vec2::new(wnd.width(), wnd.height());
 
         // convert screen position [0..resolution] to ndc [-1..1] (gpu coordinates)
         let ndc = (screen_pos / window_size) * 2.0 - Vec2::ONE;
